- moves host configuration outside of program to facilitate testability
